% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_utils.R
\name{apply_time_series_transform}
\alias{apply_time_series_transform}
\title{Apply Transformation to Time Series}
\usage{
apply_time_series_transform(series1, series2 = NULL, transform_fn, ...)
}
\arguments{
\item{series1}{First series}

\item{series2}{Second series (optional)}

\item{transform_fn}{Transformation function}

\item{...}{Additional arguments to transform_fn}
}
\value{
Transformed series
}
\description{
Apply a transformation function to aligned time series
}
\keyword{internal}
