% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{assert_bad_argument_ok}
\alias{assert_bad_argument_ok}
\title{Assert or Signal a Bad Argument Error}
\usage{
assert_bad_argument_ok(ok, message, arg = NULL)
}
\arguments{
\item{message}{Error message string}

\item{arg}{Argument name (optional)}

\item{condition}{Logical; if FALSE, signals the error}

\item{call}{The call (default NULL)}
}
\description{
Assert or Signal a Bad Argument Error
}
\keyword{internal}
