% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{assert_dimension_ok}
\alias{assert_dimension_ok}
\title{Assert or Signal a Dimension Mismatch Error}
\usage{
assert_dimension_ok(ok, message)
}
\arguments{
\item{message}{Error message string}

\item{condition}{Logical; if FALSE, signals the error}

\item{call}{The call (default NULL)}
}
\description{
Assert or Signal a Dimension Mismatch Error
}
\keyword{internal}
