% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{assert_insufficient_data_ok}
\alias{assert_insufficient_data_ok}
\title{Assert or Signal an Insufficient Data Error}
\usage{
assert_insufficient_data_ok(ok, message)
}
\arguments{
\item{message}{Error message string}

\item{condition}{Logical; if FALSE, signals the error}

\item{call}{The call (default NULL)}
}
\description{
Assert or Signal an Insufficient Data Error
}
\keyword{internal}
