% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_moments_unified.R
\name{compute_quadratic_coefficients}
\alias{compute_quadratic_coefficients}
\title{Compute Quadratic Coefficients}
\usage{
compute_quadratic_coefficients(moments, tau)
}
\arguments{
\item{moments}{List of moments from compute_gamma_moments_unified}

\item{tau}{Quantile parameter between 0 and 1}
}
\value{
List containing:
\describe{
\item{a}{Coefficient of gamma_1^2}
\item{b}{Coefficient of gamma_1}
\item{c}{Constant term}
\item{discriminant}{b^2 - 4ac}
}
}
\description{
Computes the coefficients a, b, c for the gamma quadratic equation
given the moments.
}
\keyword{internal}
