% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart_optimization.R
\name{generate_starting_points}
\alias{generate_starting_points}
\title{Generate Random Starting Points}
\usage{
generate_starting_points(
  n_starts,
  n_params,
  bounds_type = "uniform",
  lower_bounds = NULL,
  upper_bounds = NULL
)
}
\arguments{
\item{n_starts}{Number of starting points to generate}

\item{n_params}{Number of parameters}

\item{bounds_type}{Type of bounds ("uniform", "normal", "tau_weights")}

\item{lower_bounds}{Vector of lower bounds}

\item{upper_bounds}{Vector of upper bounds}
}
\value{
Matrix of starting points (n_starts x n_params)
}
\description{
Generates random starting points for multi-start optimization with
appropriate distributions for academic research parameters.
}
\keyword{internal}
