% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say Hello}
\usage{
hello(name = "World")
}
\arguments{
\item{name}{Character string with the name to greet}
}
\value{
NULL (invisibly)
}
\description{
This function prints a greeting message
}
\examples{
hello("World")
}
