% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_moments_methods.R
\name{hetid_moments_methods}
\alias{hetid_moments_methods}
\title{Methods and Assertions for hetid_moments Objects}
\description{
Internal class assertion and print method for moment containers created by
\code{\link{compute_identification_moments}}.
}
\keyword{internal}
