% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_tau0_fit_methods.R
\name{hetid_tau0_fit_methods}
\alias{hetid_tau0_fit_methods}
\title{Methods and Assertions for hetid_tau0_fit Objects}
\description{
Methods and Assertions for hetid_tau0_fit Objects
}
\keyword{internal}
