% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_theta_all_maturities.R
\name{optimize_theta_all_maturities}
\alias{optimize_theta_all_maturities}
\title{Run Theta Identification for Multiple Maturities}
\usage{
optimize_theta_all_maturities(maturities_to_test, n_pcs_options = 4, ...)
}
\arguments{
\item{maturities_to_test}{Vector of maturities to test}

\item{n_pcs_options}{Vector of PC counts to test (default 4)}

\item{...}{Additional arguments passed to optimize_theta_identification}
}
\value{
Data frame with results for all combinations
}
\description{
Convenience function to run optimize_theta_identification for multiple
maturities and/or different numbers of principal components extracted from
financial asset returns.
}
