% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_solver_base.R
\name{prepare_linear_comb_output}
\alias{prepare_linear_comb_output}
\title{Prepare Linear Combination Output}
\usage{
prepare_linear_comb_output(linear_comb, return_df, dates, n)
}
\arguments{
\item{linear_comb}{Linear combination series}

\item{return_df}{Whether to return data frame}

\item{dates}{Optional dates vector}

\item{n}{Total number of observations}
}
\value{
Either the series or a data frame
}
\description{
Common logic for preparing linear combination output in different formats
}
\keyword{internal}
