% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_log_projection_prep.R
\name{print.hetid_log_projection_prep}
\alias{print.hetid_log_projection_prep}
\title{Print a Prepared Log Projection}
\usage{
\method{print}{hetid_log_projection_prep}(x, ...)
}
\arguments{
\item{x}{A \code{hetid_log_projection_prep} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Prepared Log Projection
}
\keyword{internal}
