% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_theta_helpers.R
\name{theta_data_prep}
\alias{theta_data_prep}
\title{Data Preparation Functions for Theta Optimization}
\description{
Internal functions for preparing data for theta optimization
}
\keyword{internal}
