% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_theta_objective.R
\name{theta_objective}
\alias{theta_objective}
\title{Objective Function for Theta Optimization}
\description{
Internal function that creates the objective function for theta optimization
}
\keyword{internal}
