% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_utils.R
\name{validate_maturity_param}
\alias{validate_maturity_param}
\title{Validate Maturity Parameter}
\usage{
validate_maturity_param(i)
}
\arguments{
\item{i}{Maturity parameter}
}
\description{
Validate Maturity Parameter
}
\keyword{internal}
