% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_utils.R
\name{validate_transform_length}
\alias{validate_transform_length}
\title{Validate Transform Output Length}
\usage{
validate_transform_length(out, n_valid)
}
\arguments{
\item{out}{Output from transform_fn}

\item{n_valid}{Number of valid (non-NA) elements}
}
\description{
Validate Transform Output Length
}
\keyword{internal}
