% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_theta_box_methods.R
\name{assert_hetid_theta_box}
\alias{assert_hetid_theta_box}
\title{Assert a hetid_theta_box Object}
\usage{
assert_hetid_theta_box(x, arg = "box")
}
\arguments{
\item{x}{Object to check}

\item{arg}{Argument name used in the error message}
}
\value{
Invisibly TRUE
}
\description{
Assert a hetid_theta_box Object
}
\keyword{internal}
