% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_utils.R
\name{assert_tabular}
\alias{assert_tabular}
\title{Assert Input is Tabular}
\usage{
assert_tabular(x, name)
}
\arguments{
\item{x}{Object to check}

\item{name}{Argument name for error message}
}
\description{
Assert Input is Tabular
}
\keyword{internal}
