% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_acm_col_mapping.R
\name{build_acm_col_mapping}
\alias{build_acm_col_mapping}
\title{Build Column Mapping for ACM Data}
\usage{
build_acm_col_mapping(data_types, maturities)
}
\arguments{
\item{data_types}{Character vector of data types}

\item{maturities}{Numeric vector of maturities}
}
\value{
List mapping new names to old names
}
\description{
Internal function to build mapping between old and new column names
}
\keyword{internal}
