% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_matrix_statistics.R
\name{compute_matrix_statistics}
\alias{compute_matrix_statistics}
\title{Compute Matrix Statistics for Heteroskedasticity Identification}
\usage{
compute_matrix_statistics(w1, w2, maturities = NULL)
}
\arguments{
\item{w1}{Numeric vector of W1 residuals from compute_w1_residuals()}

\item{w2}{Matrix of W2 residuals (T x I) from compute_w2_residuals()}

\item{maturities}{Vector of maturity indices to compute statistics for.
Default is all columns of w2.}
}
\value{
A list containing:
\describe{
\item{s_i_1}{List of vectors, each element i is S_i^(1) (length I)}
\item{s_i_2}{List of matrices, each element i is S_i^(2) (I x I)}
}
}
\description{
Computes matrix statistics S_i^(1) and S_i^(2) for each maturity i.
}
\details{
First computes the matrix:
\deqn{W_2^{\circ i} = \text{diag}(W_2^{(i)}) W_2}

Then for each maturity i:
\deqn{\hat{S}_i^{(1)} = \frac{1}{T} (W_1 \odot W_2^{(i)})^T W_2^{\circ i}}
\deqn{\hat{S}_i^{(2)} = \frac{1}{T} (W_2^{\circ i})^T W_2^{\circ i}}

where \eqn{\odot} denotes the Hadamard (elementwise) product.
}
\examples{
set.seed(42)
T_obs <- 100
I <- 4
w1 <- rnorm(T_obs)
w2 <- matrix(rnorm(T_obs * I), nrow = T_obs, ncol = I)

mat_stats <- compute_matrix_statistics(w1, w2)
mat_stats$s_i_1[[1]]
mat_stats$s_i_2[[1]]
}
