% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scalar_statistics.R
\name{compute_scalar_statistics}
\alias{compute_scalar_statistics}
\title{Compute Scalar Statistics for Heteroskedasticity Identification}
\usage{
compute_scalar_statistics(w1, w2, maturities = NULL)
}
\arguments{
\item{w1}{Numeric vector of W1 residuals from compute_w1_residuals()}

\item{w2}{Matrix of W2 residuals (T x I) from compute_w2_residuals()}

\item{maturities}{Vector of maturity indices to compute statistics for.
Default is all columns of w2.}
}
\value{
A list containing:
\describe{
\item{s_i_0}{Named vector of S_i^(0) values for each maturity}
\item{sigma_i_sq}{Named vector of sigma_i^2 values for each maturity}
}
}
\description{
Computes scalar statistics S_i^(0) and sigma_i^2 for each maturity i.
}
\details{
For each maturity i, computes:
\deqn{\hat{S}_i^{(0)} = \frac{1}{T} \| W_1 \odot W_2^{(i)} \|_2^2}
\deqn{\hat{\sigma}_i^2 = \frac{1}{T} \| (W_2^{(i)})^{\odot 2} \|_2^2 -
  \left(\frac{1}{T} \| W_2^{(i)} \|_2^2 \right)^2}

where \eqn{\odot} denotes the Hadamard (elementwise) product and
\eqn{W_2^{(i)}} is the i-th column of W2.
}
\examples{
set.seed(42)
T_obs <- 100
I <- 4
w1 <- rnorm(T_obs)
w2 <- matrix(rnorm(T_obs * I), nrow = T_obs, ncol = I)

scalar_stats <- compute_scalar_statistics(w1, w2)
scalar_stats$s_i_0
scalar_stats$sigma_i_sq
}
