% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_hat_utils.R
\name{compute_time_series_news}
\alias{compute_time_series_news}
\title{Compute Time Series News}
\usage{
compute_time_series_news(current_series, future_series)
}
\arguments{
\item{current_series}{Current period series}

\item{future_series}{Future period series}
}
\value{
News series
}
\description{
Generic function to compute news as difference between future and current values
}
\keyword{internal}
