% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vector_statistics.R
\name{compute_vector_statistics}
\alias{compute_vector_statistics}
\title{Compute Vector Statistics for Heteroskedasticity Identification}
\usage{
compute_vector_statistics(w1, w2, pcs, maturities = NULL)
}
\arguments{
\item{w1}{Numeric vector of W1 residuals from compute_w1_residuals()}

\item{w2}{Matrix of W2 residuals (T x I) from compute_w2_residuals()}

\item{pcs}{Matrix of principal components (T x J)}

\item{maturities}{Vector of maturity indices to compute statistics for.
Default is all columns of w2.}
}
\value{
A list containing:
\describe{
\item{r_i_0}{Matrix (J x I) where column i contains R_i^(0)}
\item{r_i_1}{List of matrices, each element i is R_i^(1) (J x I)}
\item{p_i_0}{Matrix (J x I) where column i contains P_i^(0)}
}
}
\description{
Computes vector statistics R_i^(0), R_i^(1), and P_i^(0) for each maturity i.
}
\details{
For each maturity i, computes:
\deqn{\hat{R}_i^{(0)} = \frac{1}{T} PC^T (W_1 \odot W_2^{(i)})}
\deqn{\hat{R}_i^{(1)} = \frac{1}{T} PC^T (W_2 \odot W_2^{(i)})}
\deqn{\hat{P}_i^{(0)} = \frac{1}{T} PC^T (W_2^{(i)})^{\odot 2}}

where \eqn{\odot} denotes the Hadamard (elementwise) product.
}
\examples{
set.seed(42)
n_obs <- 100
I <- 4
J <- 3
w1 <- rnorm(n_obs)
w2 <- matrix(rnorm(n_obs * I), nrow = n_obs, ncol = I)
pcs <- matrix(rnorm(n_obs * J), nrow = n_obs, ncol = J)

vec_stats <- compute_vector_statistics(w1, w2, pcs)
vec_stats$r_i_0
vec_stats$r_i_1[[1]]
}
