% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_factory.R
\name{constraint_factory}
\alias{constraint_factory}
\title{Constraint Factory for Quadratic Form Evaluation}
\description{
Factory function for precomputing quadratic form constants
}
\keyword{internal}
