% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_quarterly.R
\name{convert_to_quarterly}
\alias{convert_to_quarterly}
\title{Convert Monthly Data to Quarterly}
\usage{
convert_to_quarterly(data)
}
\arguments{
\item{data}{Data frame with a date column}
}
\value{
Data frame with quarterly observations
}
\description{
Internal function to convert monthly data to quarterly by keeping
the last observation of each quarter.
}
\keyword{internal}
