% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acm_github.R
\name{download_acm_github}
\alias{download_acm_github}
\title{Download and Verify the ACM Data from the GitHub Release}
\usage{
download_acm_github(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical, suppress progress output}
}
\value{
Invisibly returns the cached file path
}
\description{
Download and Verify the ACM Data from the GitHub Release
}
\keyword{internal}
