% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_paths.R
\name{get_data_file_path}
\alias{get_data_file_path}
\title{Get Data File Path}
\usage{
get_data_file_path(filename)
}
\arguments{
\item{filename}{Character string with filename (including extension)}
}
\value{
Character string with full path to data file
}
\description{
Constructs full path to a data file in the package data directory.
Provides consistent path resolution.
}
\keyword{internal}
