% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_utils.R
\name{get_pc_column_names}
\alias{get_pc_column_names}
\title{Build PC Column Names}
\usage{
get_pc_column_names(n_pcs)
}
\arguments{
\item{n_pcs}{Number of principal components}
}
\value{
Character vector, e.g. c("pc1", "pc2", ...)
}
\description{
Build PC Column Names
}
\keyword{internal}
