% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvey_solver.R
\name{harvey_info}
\alias{harvey_info}
\title{Observed Information of the Harvey Criterion}
\usage{
harvey_info(theta, y, x_mat)
}
\arguments{
\item{theta}{Numeric coefficient vector of length \code{ncol(x_mat)}}

\item{y}{Numeric nonnegative response}

\item{x_mat}{Numeric design matrix, intercept column included}
}
\value{
Numeric \code{ncol(x_mat)} square matrix
}
\description{
\eqn{0.5 X' diag(r) X}, the Hessian of
\eqn{Q = 0.5 (\sum_t \eta_t + \sum_t r_t)}. The expected information
\eqn{0.5 X'X} is a trivial expression and gets no helper.
}
\keyword{internal}
