% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid-package.R
\docType{package}
\name{hetid-package}
\alias{hetid}
\alias{hetid-package}
\title{hetid: Identification Through Heteroskedasticity for VFCI}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Applies identification through heteroskedasticity from Lewbel (2012) for triangular systems to estimate the VFCI (volatility financial conditions index) by Tobias Adrian, Matthew DeHaven, Fernando Duarte and Tara Iyer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernando-duarte/hetid}
  \item Report bugs at \url{https://github.com/fernando-duarte/hetid/issues}
}

}
\author{
\strong{Maintainer}: Fernando Duarte \email{fernando_duarte@brown.edu} (\href{https://orcid.org/0000-0002-8495-6936}{ORCID})

}
\keyword{internal}
