% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid-package.R
\docType{package}
\name{hetid-package}
\alias{hetid}
\alias{hetid-package}
\title{hetid: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernando-duarte/hetid}
  \item Report bugs at \url{https://github.com/fernando-duarte/hetid/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
