% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_log_variance_fit_methods.R
\name{hetid_log_variance_fit_methods}
\alias{hetid_log_variance_fit_methods}
\title{Methods and Assertions for hetid_log_variance_fit Objects}
\description{
Methods and Assertions for hetid_log_variance_fit Objects
}
\keyword{internal}
