% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_moments_methods.R
\name{hetid_moments_methods}
\alias{hetid_moments_methods}
\title{Methods and Assertions for hetid_moments Objects}
\description{
Methods and Assertions for hetid_moments Objects
}
\keyword{internal}
