% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_term_premia.R
\name{load_term_premia}
\alias{load_term_premia}
\title{Load ACM Term Premia Data}
\usage{
load_term_premia(auto_download = FALSE)
}
\arguments{
\item{auto_download}{Logical. If TRUE and data doesn't exist, automatically
downloads it. Default is FALSE.}
}
\value{
A data frame containing the term premia data, or NULL if data is not available.
}
\description{
Loads the ACM term premia data from the package's data directory.
If the data doesn't exist, prompts the user to download it first.
}
\examples{
\dontrun{
# Load the data (will prompt to download if not available)
term_premia <- load_term_premia()

# Automatically download if not available
term_premia <- load_term_premia(auto_download = TRUE)
}

}
\seealso{
\code{\link{download_term_premia}} for downloading the data
}
