% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_log_variance_fit_methods.R
\name{print.hetid_log_variance_fit}
\alias{print.hetid_log_variance_fit}
\title{Print a hetid_log_variance_fit Object}
\usage{
\method{print}{hetid_log_variance_fit}(x, ...)
}
\arguments{
\item{x}{A \code{hetid_log_variance_fit} object}

\item{...}{Unused, for method consistency}
}
\value{
\code{x}, invisibly
}
\description{
Print a hetid_log_variance_fit Object
}
