% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_theta_box_methods.R
\name{print.hetid_theta_box}
\alias{print.hetid_theta_box}
\title{Print a hetid_theta_box Object}
\usage{
\method{print}{hetid_theta_box}(x, ...)
}
\arguments{
\item{x}{A \code{hetid_theta_box} object}

\item{...}{Unused, for method consistency}
}
\value{
\code{x}, invisibly
}
\description{
Print a hetid_theta_box Object
}
\examples{
set.seed(42)
n_obs <- 200
x <- cbind(x1 = rnorm(n_obs), x2 = rnorm(n_obs))
z <- rnorm(n_obs)
e2 <- sqrt(exp(0.5 + 0.9 * z)) * matrix(rnorm(n_obs * 2), n_obs, 2)
y2 <- x \%*\% matrix(c(1, 0.5, -0.3, 0.7), 2, 2) + e2
colnames(y2) <- c("news1", "news2")
y1 <- drop(0.3 + x \%*\% c(0.2, -0.1) + y2 \%*\% c(0.8, -0.5) + rnorm(n_obs))

fit <- compute_tau0_system(y1, y2, x, z)
box <- compute_identified_set_box(fit, tau = 0.05)
print(box)
box$beta1_bounds
}
