% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_bad_argument}
\alias{stop_bad_argument}
\title{Signal a Bad Argument Error}
\usage{
stop_bad_argument(message, arg = NULL, call = NULL)
}
\arguments{
\item{message}{Error message string}

\item{arg}{Argument name (optional)}

\item{call}{The call (default NULL)}
}
\value{
Never returns; signals a \code{hetid_error_bad_argument} condition
}
\description{
Signal a Bad Argument Error
}
\keyword{internal}
