% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_dimension_mismatch}
\alias{stop_dimension_mismatch}
\title{Signal a Dimension Mismatch Error}
\usage{
stop_dimension_mismatch(message, call = NULL)
}
\arguments{
\item{message}{Error message string}

\item{call}{The call (default NULL)}
}
\description{
Signal a Dimension Mismatch Error
}
\keyword{internal}
