% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_hetid}
\alias{stop_hetid}
\title{Signal a Generic hetid Error}
\usage{
stop_hetid(message, call = NULL)
}
\arguments{
\item{message}{Error message string}

\item{call}{The call (default NULL)}
}
\description{
Signal a Generic hetid Error
}
\keyword{internal}
