% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetid_theta_box.R
\name{validate_hetid_theta_box}
\alias{validate_hetid_theta_box}
\title{Validate a hetid_theta_box Object}
\usage{
validate_hetid_theta_box(x)
}
\arguments{
\item{x}{A \code{hetid_theta_box} object}
}
\value{
\code{x}, invisibly
}
\description{
Validate a hetid_theta_box Object
}
\keyword{internal}
