% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_w2_inputs.R
\name{validate_w2_inputs}
\alias{validate_w2_inputs}
\title{Validate and Convert W2 Input Data}
\usage{
validate_w2_inputs(yields, term_premia, maturities)
}
\arguments{
\item{yields}{Yields data (data frame or matrix)}

\item{term_premia}{Term premia data (data frame or matrix)}

\item{maturities}{Vector of maturities}
}
\value{
List with converted data frames and validated maturities
}
\description{
Internal function to validate and convert yields and term_premia inputs
}
\keyword{internal}
